requirements.txt
alembic==1.4.2
aniso8601==8.0.0
astroid==2.3.3
cffi==1.14.0
click==7.1.1
cryptography==2.9.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
pycparser==2.20
pylint==2.4.4
PyMySQL==0.9.3
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.16
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.11.2
